feature: DBScan Version Similar to Brute
I'm introducing a DBScan Version of the deduplication process under step 2 in the file name dedupe_dbscan_brute.py. It has strong similarities to the brute force method (although definitely not the same). It is way faster and the two produce slightly different results but very similar. This version is certainly suggested over the plain brute force method.